move all interfaces from devtoolsMiddleware.d.ts, fix devtoolsMiddleware import error
